синхронізація + допрацювання сторінки операції + футер

+++ Development/Beck/cto/src/main/java/com/degree/cto/controllers/TransactionsInfoController.java
@@ -1,10 +1,10 @@
 package com.degree.cto.controllers;
 
 import com.degree.cto.dtos.TransactionsInfoDTO;
-import com.degree.cto.dtos.UsersDTO;
 import com.degree.cto.logic.Log.LogService;
 import com.degree.cto.repositorys.TransactionsInfoRepository;
 import com.degree.cto.repositorys.UsersRepository;
+import com.degree.cto.security.SecurityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,6 +12,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+
+import javax.servlet.http.HttpServletRequest;
 
 @Controller
 public class TransactionsInfoController {
@@ -25,8 +27,11 @@ public class TransactionsInfoController {
     @Autowired
     private LogService logService;
 
+    @Autowired
+    private SecurityService securityService;
+
     @GetMapping("/transactions/{numberTransactions}")
-    public String transactions (@PathVariable(value = "numberTransactions") long numberTransactions, Model model) {
+    public String transactions(@PathVariable(value = "numberTransactions") long numberTransactions, Model model, HttpServletRequest request) {
         TransactionsInfoDTO transactionsInfoDTO = transactionsInfoRepository.findByNumber(numberTransactions);
         model.addAttribute("Number", transactionsInfoDTO.getNumber());
         model.addAttribute("Type", transactionsInfoDTO.getType());
@@ -38,23 +43,36 @@ public class TransactionsInfoController {
         model.addAttribute("UsernameCreator", transactionsInfoDTO.getUsernameCreator());
         model.addAttribute("CreatorRole", usersRepository.findByPersonalIndent(transactionsInfoDTO.getUsernameCreator()).getRole());
         model.addAttribute("Date", transactionsInfoDTO.getDate());
+        if (securityService.customAccess("/accountant", request.getUserPrincipal().getName(), "Бухгалтер", "Директор") != "redirect:/personal-page") {
+            model.addAttribute("EditAccess", "yes");
+        } else {
+            model.addAttribute("EditAccess", "no");
+        }
         return "transactions-info";
     }
 
     @PostMapping("/transactions/{numberTransactions}/edit")
-    public String transactionsEdit (@PathVariable(value = "numberTransactions") long numberTransactions, @ModelAttribute(value = "transactionsInfoDTO") TransactionsInfoDTO transactionsInfoDTO) {
-        transactionsInfoDTO.setId(transactionsInfoRepository.findByNumber(numberTransactions).getId());
-        transactionsInfoDTO.setNumber(numberTransactions);
-        transactionsInfoRepository.save(transactionsInfoDTO);
-        logService.addLog("log", "Виплати", "Змінено інформацію про виплату", "Виплата №" + numberTransactions + " працівником:@" + transactionsInfoRepository.findByNumber(numberTransactions).getUsernameCreator());
-        return "redirect:/transactions/" + numberTransactions;
+    public String transactionsEdit(@PathVariable(value = "numberTransactions") long numberTransactions, @ModelAttribute(value = "transactionsInfoDTO") TransactionsInfoDTO transactionsInfoDTO, HttpServletRequest request) {
+        if (securityService.customAccess("/accountant", request.getUserPrincipal().getName(), "Бухгалтер", "Директор") != "redirect:/personal-page") {
+            transactionsInfoDTO.setId(transactionsInfoRepository.findByNumber(numberTransactions).getId());
+            transactionsInfoDTO.setNumber(numberTransactions);
+            transactionsInfoRepository.save(transactionsInfoDTO);
+            logService.addLog("log", "Виплати", "Змінено інформацію про виплату", "Виплата №" + numberTransactions + " працівником:@" + transactionsInfoRepository.findByNumber(numberTransactions).getUsernameCreator());
+            return "redirect:/transactions/" + numberTransactions;
+        } else {
+            return "redirect:/personal-page";
+        }
     }
 
     @GetMapping("/transactions/{numberTransactions}/dell")
-    public String transactionsDell (@PathVariable(value = "numberTransactions") long numberTransactions) {
-        logService.addLog("log", "Виплати", "Видалено інформацію про виплату", "Виплата №" + numberTransactions + " працівником:@" + transactionsInfoRepository.findByNumber(numberTransactions).getUsernameCreator());
-        TransactionsInfoDTO transactionsInfoDTO = transactionsInfoRepository.findByNumber(numberTransactions);
-        transactionsInfoRepository.delete(transactionsInfoDTO);
-        return "redirect:/accountant/";
+    public String transactionsDell(@PathVariable(value = "numberTransactions") long numberTransactions, HttpServletRequest request) {
+        if (securityService.customAccess("/accountant", request.getUserPrincipal().getName(), "Бухгалтер", "Директор") != "redirect:/personal-page") {
+            logService.addLog("log", "Виплати", "Видалено інформацію про виплату", "Виплата №" + numberTransactions + " працівником:@" + transactionsInfoRepository.findByNumber(numberTransactions).getUsernameCreator());
+            TransactionsInfoDTO transactionsInfoDTO = transactionsInfoRepository.findByNumber(numberTransactions);
+            transactionsInfoRepository.delete(transactionsInfoDTO);
+            return "redirect:/accountant/";
+        } else {
+            return "redirect:/personal-page";
+        }
     }
 }
